+++ cuppa/toolchains/gcc.py
@@ -620,6 +620,10 @@ class Gcc(object):
         return []
 
 
+    def disable_error_limit_flags( self, env ):
+        return [ '-fmax-errors=0' ]
+
+
     def module_bmi_path( self, env, module_name ):
         from cuppa.toolchains.cxx_modules_support import named_bmi_path
         return named_bmi_path( env, module_name, '.gcm' )